Add files via upload
The Project Files folder contains two zip files. First zip file in the CLI file is the project cli which requires arduino cli and edge impulse cli to be installed onto the micro-controller in order to work. The second zip file in the Code file is the arduino code which can be deployed using the arduino ide onto the micro-controller.